Reverte "mostra mensagem se cpf e token nao forem consistentes #404"
This reverts commit fd3addf40906efa5d3939fc9035f1d41907ac1ca.

Conflicts:
	src/main/java/br/gov/servicos/editor/usuarios/RecuperacaoSenhaValidator.java
	src/main/java/br/gov/servicos/editor/usuarios/TokenCpfValido.java
	src/test/java/br/gov/servicos/editor/usuarios/RecuperacaoSenhaServiceTest.java

+++ src/main/java/br/gov/servicos/editor/usuarios/RecuperacaoSenhaService.java
@@ -25,6 +25,9 @@ public class RecuperacaoSenhaService {
     @Autowired
     private UsuarioRepository usuarioRepository;
 
+    @Autowired
+    private RecuperacaoSenhaValidator validator;
+
     private Clock clock = Clock.systemUTC();
 
     public String gerarTokenParaUsuario(String usuarioId) {
@@ -38,13 +41,18 @@ public class RecuperacaoSenhaService {
         return token;
     }
 
-    public void trocarSenha(FormularioRecuperarSenha formulario) {
+    public boolean trocarSenha(FormularioRecuperarSenha formulario) {
         Long usuarioId = formulario.getUsuarioId();
         TokenRecuperacaoSenha token = repository.findByUsuarioId(usuarioId);
         Usuario usuario = token.getUsuario();
 
-        usuarioRepository.save(usuario.withSenha(passwordEncoder.encode(formulario.getCamposSenha().getSenha())));
-        repository.delete(token.getId());
+        if(validator.isValid(formulario, token)) {
+            usuarioRepository.save(usuario.withSenha(passwordEncoder.encode(formulario.getCamposSenha().getSenha())));
+            repository.delete(token.getId());
+            return true;
+        } else {
+            return false;
+        }
     }
 
     public void falhaNaVerificacao(Long usuarioId) {
